@@ -14,61 +14,63 @@ Status: Implementation in progress
 """
 
 from typing import List, Dict, Any, Optional
-from rag.core.base import Embedder, Document
+from core.base import Embedder, Document
 
 
 class HuggingFaceEmbedder(Embedder):
     """HuggingFace-based embedder for generating document embeddings.
-    
+
     This component will integrate with the HuggingFace transformers ecosystem
     to provide high-quality embeddings using state-of-the-art models.
-    
+
     Status: Coming Soon
     """
-    
-    def __init__(self, name: str = "HuggingFaceEmbedder", config: Dict[str, Any] = None):
+
+    def __init__(
+        self, name: str = "HuggingFaceEmbedder", config: Dict[str, Any] = None
+    ):
         super().__init__(name, config)
-        
+
         # Configuration will include:
         # - model_name: HuggingFace model identifier
         # - device: CPU/GPU device selection
         # - max_length: Maximum sequence length
         # - batch_size: Processing batch size
         # - trust_remote_code: Whether to trust remote code
-        
+
         # Placeholder for future implementation
         raise NotImplementedError(
             "HuggingFaceEmbedder is not yet implemented. "
             "Coming soon in a future release! "
             "For now, please use OllamaEmbedder or OpenAIEmbedder."
         )
-    
+
     def embed_documents(self, documents: List[Document]) -> List[Document]:
         """Generate embeddings for a batch of documents.
-        
+
         This will use HuggingFace transformers to generate high-quality
         embeddings from the document content.
-        
+
         Args:
             documents: List of documents to embed
-            
+
         Returns:
             Documents with embeddings added
-            
+
         Raises:
             NotImplementedError: This embedder is not yet implemented
         """
         raise NotImplementedError("HuggingFaceEmbedder coming soon!")
-    
+
     def embed_query(self, query: str) -> List[float]:
         """Generate embedding for a single query string.
-        
+
         Args:
             query: Query string to embed
-            
+
         Returns:
             Query embedding vector
-            
+
         Raises:
             NotImplementedError: This embedder is not yet implemented
         """
