@@ -14,21 +14,21 @@ Status: Implementation in progress
 """
 
 from typing import List, Dict, Any, Optional
-from rag.core.base import Embedder, Document
+from core.base import Embedder, Document
 
 
 class OpenAIEmbedder(Embedder):
     """OpenAI API-based embedder for generating document embeddings.
-    
+
     This component will integrate with OpenAI's embedding API to provide
     state-of-the-art embeddings using their latest models.
-    
+
     Status: Coming Soon
     """
-    
+
     def __init__(self, name: str = "OpenAIEmbedder", config: Dict[str, Any] = None):
         super().__init__(name, config)
-        
+
         # Configuration will include:
         # - api_key: OpenAI API key
         # - model: Model name (text-embedding-3-small, text-embedding-3-large)
@@ -36,40 +36,40 @@ class OpenAIEmbedder(Embedder):
         # - batch_size: API batch size
         # - max_retries: Maximum retry attempts
         # - timeout: Request timeout
-        
+
         # Placeholder for future implementation
         raise NotImplementedError(
             "OpenAIEmbedder is not yet implemented. "
             "Coming soon in a future release! "
             "For now, please use OllamaEmbedder."
         )
-    
+
     def embed_documents(self, documents: List[Document]) -> List[Document]:
         """Generate embeddings for a batch of documents using OpenAI API.
-        
+
         This will use OpenAI's text-embedding API to generate high-quality
         embeddings with proper rate limiting and error handling.
-        
+
         Args:
             documents: List of documents to embed
-            
+
         Returns:
             Documents with embeddings added
-            
+
         Raises:
             NotImplementedError: This embedder is not yet implemented
         """
         raise NotImplementedError("OpenAIEmbedder coming soon!")
-    
+
     def embed_query(self, query: str) -> List[float]:
         """Generate embedding for a single query string.
-        
+
         Args:
             query: Query string to embed
-            
+
         Returns:
             Query embedding vector
-            
+
         Raises:
             NotImplementedError: This embedder is not yet implemented
         """
