@@ -14,21 +14,23 @@ Status: Implementation in progress
 """
 
 from typing import List, Dict, Any, Optional
-from rag.core.base import Embedder, Document
+from core.base import Embedder, Document
 
 
 class SentenceTransformerEmbedder(Embedder):
     """Sentence Transformers-based embedder for generating document embeddings.
-    
+
     This component will integrate with the sentence-transformers library
     to provide local, high-quality embeddings without API dependencies.
-    
+
     Status: Coming Soon
     """
-    
-    def __init__(self, name: str = "SentenceTransformerEmbedder", config: Dict[str, Any] = None):
+
+    def __init__(
+        self, name: str = "SentenceTransformerEmbedder", config: Dict[str, Any] = None
+    ):
         super().__init__(name, config)
-        
+
         # Configuration will include:
         # - model_name: Sentence transformer model name
         # - device: CPU/GPU device selection
@@ -36,40 +38,40 @@ class SentenceTransformerEmbedder(Embedder):
         # - batch_size: Processing batch size
         # - show_progress_bar: Display progress during encoding
         # - convert_to_tensor: Return tensors instead of numpy
-        
+
         # Placeholder for future implementation
         raise NotImplementedError(
             "SentenceTransformerEmbedder is not yet implemented. "
             "Coming soon in a future release! "
             "For now, please use OllamaEmbedder."
         )
-    
+
     def embed_documents(self, documents: List[Document]) -> List[Document]:
         """Generate embeddings for a batch of documents using sentence transformers.
-        
+
         This will use the sentence-transformers library to generate embeddings
         locally without any API calls.
-        
+
         Args:
             documents: List of documents to embed
-            
+
         Returns:
             Documents with embeddings added
-            
+
         Raises:
             NotImplementedError: This embedder is not yet implemented
         """
         raise NotImplementedError("SentenceTransformerEmbedder coming soon!")
-    
+
     def embed_query(self, query: str) -> List[float]:
         """Generate embedding for a single query string.
-        
+
         Args:
             query: Query string to embed
-            
+
         Returns:
             Query embedding vector
-            
+
         Raises:
             NotImplementedError: This embedder is not yet implemented
         """
