@@ -3,7 +3,7 @@
 from pathlib import Path
 from typing import Dict, Any, Optional, List
 import logging
-from rag.components.parsers.base.base_parser import BaseParser, ParserConfig
+from components.parsers.base.base_parser import BaseParser, ParserConfig
 
 logger = logging.getLogger(__name__)
 
@@ -40,7 +40,7 @@ class PDFParser_LlamaIndex(BaseParser):
                 "pypdf2_fallback",
             ],
         )
-    
+
     def _load_metadata(self) -> ParserConfig:
         """Load parser metadata."""
         return ParserConfig(
@@ -49,74 +49,88 @@ class PDFParser_LlamaIndex(BaseParser):
             version="1.0.0",
             supported_extensions=[".pdf"],
             mime_types=["application/pdf"],
-            capabilities=["text_extraction", "metadata_extraction", "table_extraction", "image_extraction"],
-            dependencies={"llama-index": ["llama-index>=0.9.0"], "PyMuPDF": ["PyMuPDF>=1.23.0"]},
+            capabilities=[
+                "text_extraction",
+                "metadata_extraction",
+                "table_extraction",
+                "image_extraction",
+            ],
+            dependencies={
+                "llama-index": ["llama-index>=0.9.0"],
+                "PyMuPDF": ["PyMuPDF>=1.23.0"],
+            },
             default_config={
                 "chunk_size": 1000,
                 "chunk_overlap": 100,
                 "chunk_strategy": "characters",
                 "extract_metadata": True,
                 "extract_images": False,
-                "extract_tables": False
-            }
+                "extract_tables": False,
+            },
         )
-    
+
     def can_parse(self, file_path: str) -> bool:
         """Check if this parser can handle the given file."""
-        return file_path.lower().endswith('.pdf')
-    
+        return file_path.lower().endswith(".pdf")
+
     def parse_blob(self, data: bytes, metadata: Dict[str, Any] = None) -> List:
         """Parse PDF from raw bytes."""
-        from rag.core.base import Document
+        from core.base import Document
         import io
         import tempfile
         import os
-        
+
         try:
             from llama_index.core import SimpleDirectoryReader
         except ImportError:
             # Fall back to PyPDF2 parser
-            from rag.components.parsers.pdf.pypdf2_parser import PDFParser_PyPDF2
+            from components.parsers.pdf.pypdf2_parser import PDFParser_PyPDF2
+
             fallback_parser = PDFParser_PyPDF2(config=self.config)
             return fallback_parser.parse_blob(data, metadata)
-        
+
         try:
             # LlamaIndex needs a file on disk, so write temporarily
-            with tempfile.NamedTemporaryFile(suffix='.pdf', delete=False) as tmp_file:
+            with tempfile.NamedTemporaryFile(suffix=".pdf", delete=False) as tmp_file:
                 tmp_file.write(data)
                 tmp_path = tmp_file.name
-            
+
             try:
                 # Use LlamaIndex to read the PDF
                 reader = SimpleDirectoryReader(input_files=[tmp_path])
                 llama_docs = reader.load_data()
-                
+
                 documents = []
-                filename = metadata.get("filename", "unknown.pdf") if metadata else "unknown.pdf"
-                
+                filename = (
+                    metadata.get("filename", "unknown.pdf")
+                    if metadata
+                    else "unknown.pdf"
+                )
+
                 for llama_doc in llama_docs:
                     doc = Document(
                         content=llama_doc.text,
                         metadata={
                             "source": filename,
                             "parser": "PDFParser_LlamaIndex",
-                            **(metadata or {})
+                            **(metadata or {}),
                         },
-                        source=filename
+                        source=filename,
                     )
                     documents.append(doc)
-                
+
                 return documents
-                
+
             finally:
                 # Clean up temp file
                 if os.path.exists(tmp_path):
                     os.unlink(tmp_path)
-                    
+
         except Exception as e:
             print(f"Error parsing PDF with LlamaIndex, falling back to PyPDF2: {e}")
             # Fall back to PyPDF2 parser
-            from rag.components.parsers.pdf.pypdf2_parser import PDFParser_PyPDF2
+            from components.parsers.pdf.pypdf2_parser import PDFParser_PyPDF2
+
             fallback_parser = PDFParser_PyPDF2(config=self.config)
             return fallback_parser.parse_blob(data, metadata)
 
@@ -126,7 +140,7 @@ class PDFParser_LlamaIndex(BaseParser):
 
     def parse(self, source: str, **kwargs):
         """Parse PDF using LlamaIndex with fallback strategies."""
-        from rag.core.base import Document, ProcessingResult
+        from core.base import Document, ProcessingResult
 
         path = Path(source)
         if not path.exists():
@@ -149,7 +163,7 @@ class PDFParser_LlamaIndex(BaseParser):
 
     def _try_strategy(self, strategy: str, path: Path):
         """Try a specific parsing strategy."""
-        from rag.core.base import Document, ProcessingResult
+        from core.base import Document, ProcessingResult
 
         try:
             if strategy == "llama_pdf_reader":
@@ -166,7 +180,7 @@ class PDFParser_LlamaIndex(BaseParser):
 
     def _parse_with_llama_pdf(self, path: Path):
         """Parse using LlamaIndex PDFReader."""
-        from rag.core.base import Document, ProcessingResult
+        from core.base import Document, ProcessingResult
 
         try:
             from llama_index.readers.file import PDFReader
@@ -254,7 +268,7 @@ class PDFParser_LlamaIndex(BaseParser):
 
     def _parse_with_llama_pymupdf(self, path: Path):
         """Parse using LlamaIndex PyMuPDFReader."""
-        from rag.core.base import Document, ProcessingResult
+        from core.base import Document, ProcessingResult
 
         try:
             from llama_index.readers.file import PyMuPDFReader
@@ -338,7 +352,7 @@ class PDFParser_LlamaIndex(BaseParser):
 
     def _parse_with_direct_pymupdf(self, path: Path):
         """Parse directly using PyMuPDF."""
-        from rag.core.base import Document, ProcessingResult
+        from core.base import Document, ProcessingResult
 
         try:
             import fitz  # PyMuPDF
@@ -424,7 +438,7 @@ class PDFParser_LlamaIndex(BaseParser):
 
     def _parse_with_pypdf2(self, path: Path):
         """Fallback to PyPDF2."""
-        from rag.core.base import Document, ProcessingResult
+        from core.base import Document, ProcessingResult
 
         try:
             import PyPDF2
