@@ -8,6 +8,6 @@ where each strategy is a self-contained component with its own configuration.
 from .base import RetrievalStrategy, RetrievalResult, HybridRetrievalStrategy
 
 # Individual strategies can be imported as needed from their subdirectories
-# Example: from rag.components.retrievers.basic_similarity.basic_similarity import BasicSimilarityStrategy
+# Example: from components.retrievers.basic_similarity.basic_similarity import BasicSimilarityStrategy
 
 __all__ = ["RetrievalStrategy", "RetrievalResult", "HybridRetrievalStrategy"]
