@@ -15,21 +15,21 @@ Status: Implementation in progress
 """
 
 from typing import List, Dict, Any, Optional
-from rag.core.base import VectorStore, Document
+from core.base import VectorStore, Document
 
 
 class QdrantStore(VectorStore):
     """Qdrant-based vector store for high-performance similarity search.
-    
+
     This component will integrate with Qdrant to provide enterprise-grade
     vector search capabilities with advanced filtering and optimization.
-    
+
     Status: Coming Soon
     """
-    
+
     def __init__(self, name: str = "QdrantStore", config: Dict[str, Any] = None):
         super().__init__(name, config)
-        
+
         # Configuration will include:
         # - url: Qdrant server URL
         # - api_key: API key for Qdrant Cloud
@@ -39,66 +39,66 @@ class QdrantStore(VectorStore):
         # - hnsw_config: HNSW index configuration
         # - quantization_config: Quantization settings
         # - replication_factor: Number of replicas
-        
+
         # Placeholder for future implementation
         raise NotImplementedError(
             "QdrantStore is not yet implemented. "
             "Coming soon in a future release! "
             "For now, please use ChromaStore."
         )
-    
+
     def add(self, documents: List[Document]) -> None:
         """Add documents to Qdrant collection.
-        
+
         This will insert documents with their embeddings and metadata
         into Qdrant with proper error handling and batching.
-        
+
         Args:
             documents: Documents to add to the collection
-            
+
         Raises:
             NotImplementedError: This store is not yet implemented
         """
         raise NotImplementedError("QdrantStore coming soon!")
-    
+
     def search(
         self,
         query_embedding: List[float],
         top_k: int = 5,
-        metadata_filter: Optional[Dict[str, Any]] = None
+        metadata_filter: Optional[Dict[str, Any]] = None,
     ) -> List[Document]:
         """Search for similar documents in Qdrant.
-        
+
         Args:
             query_embedding: Query vector
             top_k: Number of results to return
             metadata_filter: Optional metadata filters
-            
+
         Returns:
             List of similar documents
-            
+
         Raises:
             NotImplementedError: This store is not yet implemented
         """
         raise NotImplementedError("QdrantStore coming soon!")
-    
+
     def delete(self, document_ids: List[str]) -> None:
         """Delete documents from Qdrant collection.
-        
+
         Args:
             document_ids: IDs of documents to delete
-            
+
         Raises:
             NotImplementedError: This store is not yet implemented
         """
         raise NotImplementedError("QdrantStore coming soon!")
-    
+
     def get_collection_info(self) -> Dict[str, Any]:
         """Get information about the Qdrant collection.
-        
+
         Returns:
             Collection statistics and configuration
-            
+
         Raises:
             NotImplementedError: This store is not yet implemented
         """
