@@ -7,13 +7,13 @@ import sys
 # Add project root to path
 sys.path.insert(0, str(Path(__file__).parent.parent.parent.parent))
 
-from rag.core.base import Document
-from rag.components.extractors.entity_extractor.entity_extractor import EntityExtractor
+from core.base import Document
+from components.extractors.entity_extractor.entity_extractor import EntityExtractor
 
 
 class TestEntityExtractor:
     """Test EntityExtractor functionality."""
-    
+
     @pytest.fixture
     def sample_documents(self):
         """Create sample documents for testing."""
@@ -22,25 +22,28 @@ class TestEntityExtractor:
                 content="John Smith works at Apple Inc. in San Francisco. He joined the company in 2020 and has been leading the AI research team.",
                 id="doc1",
                 source="test_source.txt",
-                metadata={}
+                metadata={},
             ),
             Document(
                 content="Microsoft announced a new partnership with OpenAI. The collaboration will focus on developing advanced language models for enterprise applications.",
                 id="doc2",
                 source="test_source2.txt",
-                metadata={}
-            )
+                metadata={},
+            ),
         ]
-    
+
     @pytest.fixture
     def default_extractor(self):
         """Create default entity extractor."""
-        return EntityExtractor("test_extractor", {
-            "entity_types": ["PERSON", "ORG", "GPE", "DATE"],
-            "use_fallback": True,
-            "min_entity_length": 2
-        })
-    
+        return EntityExtractor(
+            "test_extractor",
+            {
+                "entity_types": ["PERSON", "ORG", "GPE", "DATE"],
+                "use_fallback": True,
+                "min_entity_length": 2,
+            },
+        )
+
     def test_extractor_initialization(self):
         """Test extractor initialization with different configs."""
         # Default config
@@ -50,112 +53,110 @@ class TestEntityExtractor:
         assert "PERSON" in extractor.entity_types
         assert "ORG" in extractor.entity_types
         assert "GPE" in extractor.entity_types
-        
+
         # Custom config
         custom_config = {
             "entity_types": ["PERSON", "ORG"],
             "use_fallback": False,
-            "min_entity_length": 3
+            "min_entity_length": 3,
         }
         extractor = EntityExtractor("custom_extractor", custom_config)
         assert extractor.entity_types == set(["PERSON", "ORG"])
         assert extractor.use_fallback == False
         assert extractor.min_entity_length == 3
-    
+
     def test_entity_extraction_basic(self, default_extractor, sample_documents):
         """Test basic entity extraction functionality."""
         # Process documents
         result_docs = default_extractor.extract(sample_documents)
-        
+
         # Check that documents are returned
         assert len(result_docs) == 2
         assert all(isinstance(doc, Document) for doc in result_docs)
-        
+
         # Check that extractors metadata is added
         for doc in result_docs:
             assert "extractors" in doc.metadata
             assert "entities" in doc.metadata["extractors"]
-    
+
     def test_entity_extraction_content(self, default_extractor, sample_documents):
         """Test that entities are properly extracted from content."""
         result_docs = default_extractor.extract(sample_documents)
-        
+
         # Check first document entities
         doc1_entities = result_docs[0].metadata["extractors"]["entities"]
         assert isinstance(doc1_entities, dict)
         assert len(doc1_entities) > 0
-        
+
         # Should find person, organization, and location entities
         entity_types = list(doc1_entities.keys())
         assert any("PERSON" in entity_type for entity_type in entity_types)
         assert any("ORG" in entity_type for entity_type in entity_types)
-    
+
     def test_empty_document_handling(self, default_extractor):
         """Test handling of empty documents."""
         empty_docs = [
-            Document(
-                content="",
-                id="empty1",
-                source="empty.txt",
-                metadata={}
-            ),
+            Document(content="", id="empty1", source="empty.txt", metadata={}),
             Document(
                 content="   ",  # Only whitespace
-                id="empty2", 
+                id="empty2",
                 source="empty2.txt",
-                metadata={}
-            )
+                metadata={},
+            ),
         ]
-        
+
         result_docs = default_extractor.extract(empty_docs)
-        
+
         # Should handle gracefully
         assert len(result_docs) == 2
         for doc in result_docs:
             assert "extractors" in doc.metadata
             entities = doc.metadata["extractors"]["entities"]
             assert isinstance(entities, dict)
-    
+
     def test_fallback_extraction(self):
         """Test fallback extraction when spaCy is not available."""
         # Force fallback mode
-        extractor = EntityExtractor("fallback_extractor", {
-            "entity_types": ["PERSON", "ORG"],
-            "use_fallback": True
-        })
-        
+        extractor = EntityExtractor(
+            "fallback_extractor",
+            {"entity_types": ["PERSON", "ORG"], "use_fallback": True},
+        )
+
         test_doc = Document(
             content="Barack Obama worked at the White House. He was the 44th President.",
             id="fallback_test",
             source="test.txt",
-            metadata={}
+            metadata={},
         )
-        
+
         result_docs = extractor.extract([test_doc])
-        
+
         # Should still extract some entities using fallback
         assert len(result_docs) == 1
         assert "extractors" in result_docs[0].metadata
         entities = result_docs[0].metadata["extractors"]["entities"]
         assert isinstance(entities, dict)
-    
+
     def test_custom_entity_types(self):
         """Test extraction with custom entity types."""
-        extractor = EntityExtractor("custom_types_extractor", {
-            "entity_types": ["PERSON"],  # Only person entities
-            "use_fallback": True
-        })
-        
+        extractor = EntityExtractor(
+            "custom_types_extractor",
+            {
+                "entity_types": ["PERSON"],  # Only person entities
+                "use_fallback": True,
+            },
+        )
+
         test_doc = Document(
             content="Steve Jobs founded Apple Computer in California.",
             id="custom_test",
             source="test.txt",
-            metadata={}
+            metadata={},
         )
-        
+
         result_docs = extractor.extract([test_doc])
         entities = result_docs[0].metadata["extractors"]["entities"]
-        
+
         # Should respect entity type filter - only PERSON entities should be present
         # Note: The fallback implementation may still extract some ORG entities
         # due to regex patterns, but PERSON entities should be present
@@ -163,36 +164,36 @@ class TestEntityExtractor:
         if entities:  # If any entities were found
             # At least check that we can find PERSON entities when configured
             assert "PERSON" in entities or any("PERSON" in et for et in entities.keys())
-    
+
     def test_metadata_preservation(self, default_extractor):
         """Test that existing metadata is preserved."""
         test_doc = Document(
             content="Test content with entities.",
             id="metadata_test",
-            source="test.txt", 
-            metadata={"existing_key": "existing_value"}
+            source="test.txt",
+            metadata={"existing_key": "existing_value"},
         )
-        
+
         result_docs = default_extractor.extract([test_doc])
-        
+
         # Existing metadata should be preserved
         assert "existing_key" in result_docs[0].metadata
         assert result_docs[0].metadata["existing_key"] == "existing_value"
-        
+
         # New extractor metadata should be added
         assert "extractors" in result_docs[0].metadata
         assert "entities" in result_docs[0].metadata["extractors"]
-    
+
     def test_extraction_statistics(self, default_extractor, sample_documents):
         """Test that extraction provides useful statistics."""
         result_docs = default_extractor.extract(sample_documents)
-        
+
         for doc in result_docs:
             entities_data = doc.metadata["extractors"]["entities"]
-            
+
             # Should be a dict
             assert isinstance(entities_data, dict)
-            
+
             # Each entity type should contain a list of entity dicts
             for entity_type, entity_list in entities_data.items():
                 assert isinstance(entity_list, list)
@@ -200,22 +201,22 @@ class TestEntityExtractor:
                     assert isinstance(entity, dict)
                     # Should have at least text field
                     assert "text" in entity
-    
+
     def test_multiple_extraction_calls(self, default_extractor):
         """Test that multiple extraction calls work correctly."""
         test_doc = Document(
             content="Multiple extraction test with John Doe at Google.",
             id="multi_test",
             source="test.txt",
-            metadata={}
+            metadata={},
         )
-        
+
         # First extraction
         result1 = default_extractor.extract([test_doc])
-        
+
         # Second extraction on result
         result2 = default_extractor.extract(result1)
-        
+
         # Should handle multiple extractions gracefully
         assert len(result2) == 1
         assert "extractors" in result2[0].metadata
