@@ -4,9 +4,9 @@ import pytest
 import tempfile
 from pathlib import Path
 
-from rag.core.blob_processor import BlobProcessor
-from rag.core.ingest_handler import IngestHandler
-from rag.core.factories import ComponentFactory
+from core.blob_processor import BlobProcessor
+from core.ingest_handler import IngestHandler
+from core.factories import ComponentFactory
 
 
 class TestCoreModules:
@@ -14,58 +14,55 @@ class TestCoreModules:
 
     def test_blob_processor_initialization(self):
         """Test BlobProcessor initialization."""
-        strategy_config = {
-            "parsers": [{"type": "TextParser_Python", "config": {}}]
-        }
+        strategy_config = {"parsers": [{"type": "TextParser_Python", "config": {}}]}
         processor = BlobProcessor(strategy_config)
-        
+
         assert processor is not None
-        assert hasattr(processor, 'process_blob')
+        assert hasattr(processor, "process_blob")
 
     def test_blob_processor_text_file(self):
         """Test processing a text blob."""
-        strategy_config = {
-            "parsers": [{"type": "TextParser_Python", "config": {}}]
-        }
+        strategy_config = {"parsers": [{"type": "TextParser_Python", "config": {}}]}
         processor = BlobProcessor(strategy_config)
-        
-        with tempfile.NamedTemporaryFile(mode='w', suffix='.txt', delete=False) as f:
+
+        with tempfile.NamedTemporaryFile(mode="w", suffix=".txt", delete=False) as f:
             f.write("Test content for blob processing")
             f.flush()
-            
+
             # Process the file as blob
-            with open(f.name, 'rb') as file:
+            with open(f.name, "rb") as file:
                 blob_data = file.read()
             result = processor.process_blob(blob_data, {"filename": f.name})
-            
+
             assert result is not None
             Path(f.name).unlink()
 
     @pytest.mark.skip(reason="Requires complex setup with running services")
     def test_ingest_handler_initialization(self):
         """Test IngestHandler initialization."""
-        with tempfile.NamedTemporaryFile(mode='w', suffix='.yaml', delete=False) as f:
+        with tempfile.NamedTemporaryFile(mode="w", suffix=".yaml", delete=False) as f:
             import yaml
+
             config = {
                 "databases": {"test_db": {"vector_store": {"type": "chroma"}}},
                 "data_processing_strategies": [
                     {"name": "test_strategy", "parsers": [], "extractors": []}
-                ]
+                ],
             }
             yaml.dump(config, f)
             f.flush()
-            
+
             handler = IngestHandler(f.name, "test_strategy", "test_db")
-            
+
             assert handler is not None
-            assert hasattr(handler, 'ingest')
-            
+            assert hasattr(handler, "ingest")
+
             Path(f.name).unlink()
 
     def test_component_factory_parser_creation(self):
         """Test creating parsers via factory."""
-        from rag.core.factories import create_parser_from_config
-        
+        from core.factories import create_parser_from_config
+
         # Test creating text parser
         config = {"type": "text", "config": {}}
         parser = create_parser_from_config(config)
@@ -73,34 +70,33 @@ class TestCoreModules:
 
     def test_component_factory_embedder_creation(self):
         """Test creating embedders via factory."""
-        from rag.core.factories import OllamaEmbedder
-        
+        from core.factories import OllamaEmbedder
+
         # Test creating ollama embedder directly
         embedder = OllamaEmbedder(config={"model": "nomic-embed-text"})
         assert embedder is not None
-        assert hasattr(embedder, 'embed')
+        assert hasattr(embedder, "embed")
 
     def test_component_factory_store_creation(self):
         """Test creating stores via factory."""
         try:
-            from rag.core.factories import ChromaStore
-            
+            from core.factories import ChromaStore
+
             with tempfile.TemporaryDirectory() as temp_dir:
                 # Test creating ChromaDB store directly
-                store = ChromaStore(config={
-                    "collection_name": "test",
-                    "persist_directory": temp_dir
-                })
+                store = ChromaStore(
+                    config={"collection_name": "test", "persist_directory": temp_dir}
+                )
                 assert store is not None
-                assert hasattr(store, 'add_documents')
+                assert hasattr(store, "add_documents")
         except ImportError:
             # ChromaDB not installed
             assert True
 
     def test_factory_error_handling(self):
         """Test factory handles unknown types."""
-        from rag.core.factories import create_parser_from_config
-        
+        from core.factories import create_parser_from_config
+
         # Should handle unknown parser type
         config = {"type": "unknown_parser_type", "config": {}}
         try:
@@ -112,24 +108,25 @@ class TestCoreModules:
     @pytest.mark.skip(reason="Requires complex setup with running services")
     def test_ingest_handler_file_processing(self):
         """Test file ingestion process."""
-        with tempfile.NamedTemporaryFile(mode='w', suffix='.yaml', delete=False) as f:
+        with tempfile.NamedTemporaryFile(mode="w", suffix=".yaml", delete=False) as f:
             import yaml
+
             config = {
                 "databases": {"test_db": {"vector_store": {"type": "chroma"}}},
                 "data_processing_strategies": [
                     {"name": "test_strategy", "parsers": [], "extractors": []}
-                ]
+                ],
             }
             yaml.dump(config, f)
             f.flush()
-            
+
             handler = IngestHandler(f.name, "test_strategy", "test_db")
             Path(f.name).unlink()
-        
-        with tempfile.NamedTemporaryFile(mode='w', suffix='.txt', delete=False) as f:
+
+        with tempfile.NamedTemporaryFile(mode="w", suffix=".txt", delete=False) as f:
             f.write("Document to ingest")
             f.flush()
-            
+
             # Test ingestion (may need configuration)
             try:
                 result = handler.ingest(f.name)
