@@ -1,3 +1,11 @@
-from .task_process_dataset import process_dataset_task
+from .task_process_dataset import (
+    process_dataset_task,
+    process_single_file_task,
+    create_dataset_processing_chain,
+)
 
-__all__ = ["process_dataset_task"]
+__all__ = [
+    "process_dataset_task",
+    "process_single_file_task",
+    "create_dataset_processing_chain",
+]
